refactor(explorer): extract IssueFocus module (#316)
Move IssueFocus + participantsOf + longestCommonAncestor + the
edge/container derivation out of App.tsx and into
state/issueFocus.ts behind one pure function:
computeIssueFocus(issue, key, data) → IssueFocus.

App.tsx's onIssueFocus handler shrinks from ~50 lines to 4.
No behaviour change; the 30-case Playwright smoke suite still
passes. Unblocks #318 (HighlightPolicy) where focus becomes one
input to a unified policy instead of a sibling of pathFinder
and minCut memos.

Unit-test framework not yet wired in the template (Playwright
e2e only) — leaving #316's "unit tests" acceptance item for a
follow-up that lands vitest at the same time as the next batch
of state modules (queries, highlightPolicy).
