+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- **Explorer: extract `IssueFocus` module from `App.tsx`** (#316): The canvas-level focus state driven by Issues-tab selections — `participantsOf`, `longestCommonAncestor`, expanded-container derivation, and the inter-participant edge set — moves into `crates/noupling-explorer/template/src/state/issueFocus.ts` behind one public function `computeIssueFocus(issue, key, data) → IssueFocus`. `App.tsx`'s `onIssueFocus` handler shrinks from ~50 inline lines to a four-line call. Pure function, no React, no setState; callers still own the side effect of scoping the view to `focus.lca`. Unit-test framework is not yet wired in the template subproject (Playwright e2e only), so verification is the existing smoke suite — all 30 specs still pass. Lays groundwork for #318 (HighlightPolicy) where the focus rule becomes one input to a unified canvas-highlight policy instead of a tangled sibling of `pathFinder` and `minCut` memos.
+
 - **Workspace split: `noupling-core` / `noupling` (cli) / `noupling-explorer`** (#229, Task 1 of #228): The single-crate repo became a Cargo workspace with three crates under `crates/`. `noupling-core` is the pure analysis library (analyzer, scanner, storage, layers/rules, settings, baseline, diff). The cli crate (package name still `noupling`, binary still `noupling`) is the only consumer that wires reporter + cli code; it depends on `noupling-core` and `noupling-explorer`. `noupling-explorer` is an empty stub crate ready for Task 2 of #228. Dependency direction is strictly one-way: `cli → explorer → core`, `cli → core`. The boundary is dogfooded in `.noupling/settings.json` via new `dependency_rules` that forbid core → cli/explorer and explorer → cli/reporter; the same boundary is enforced at compile time by the workspace layout. The `analyzer::AuditResultBuilder` test helper is now exposed via a `test-utils` cargo feature on `noupling-core`, enabled as a dev-dependency feature on the cli crate. CI (`cargo check/clippy/test`) switched to `--workspace` and `cargo fmt` switched to `--all`. `scripts/release.sh` likewise. Binary surface is unchanged (`noupling --help` byte-identical, all integration tests pass without modification). **Coordinated follow-ups required before the next release tag:** Homebrew tap formula must be updated to `cargo install --path crates/noupling-cli --bin noupling`; crates.io publish flow needs an explicit decision (publish core/explorer/cli in dep order, or drop crates.io publishing) — see #229 for the checklist.
 
 ### Added
