+++ crates/noupling-explorer/template/src/App.tsx
@@ -10,24 +10,7 @@ import { ScoreDialog } from "./components/ScoreDialog";
 import type { Issue } from "./components/SidePanel";
 import { useExplorerStore, useNodeFilter, inScope } from "./state/explorerState";
 import { shortestPath, pathEdges, minCutEdges } from "./state/paths";
-
-interface IssueFocus {
-  /** Stable key matching the IssuesTab's button data-issue-key. */
-  key: string;
-  /** Lowest common ancestor of the participants — what scope we drilled to. */
-  lca: string;
-  /** Concrete edges between participants — extra-prominent on the canvas. */
-  edges: Set<string>;
-  /** Immediate-children-of-LCA containers that hold at least one
-   *  participant. These render *expanded* on the canvas (file children
-   *  shown inline) instead of as a single container card. #275
-   *  follow-up. */
-  expandedContainers: Set<string>;
-  /** The actual participant file ids that live under those containers
-   *  — the only files to surface in the expanded render so the canvas
-   *  doesn't hairball on many-file packages. */
-  participantFiles: Set<string>;
-}
+import { computeIssueFocus, type IssueFocus } from "./state/issueFocus";
 
 export interface AppProps {
   data: DataContract;
@@ -99,47 +82,9 @@ export function App({ data }: AppProps) {
       setIssueFocus(null);
       return;
     }
-    // Collect participants per issue kind so we can compute LCA + edge
-    // highlights. Falls back to single-subject when an issue family
-    // doesn't enumerate multiple files.
-    const participants = participantsOf(issue, data);
-    const lca = longestCommonAncestor(participants);
-    const edges = new Set<string>();
-    if (participants.length >= 2) {
-      // Highlight all directed edges between any pair of participants.
-      const set = new Set(participants);
-      for (const e of data.edges) {
-        if (set.has(e.from) && set.has(e.to)) {
-          edges.add(`${e.from}→${e.to}`);
-        }
-      }
-    }
-    // Determine which *immediate child of LCA* each participant lives
-    // under — those are the containers we'll expand inline on the
-    // canvas so the offending edges render at file level. Participants
-    // that ARE the immediate child stay as themselves.
-    const expandedContainers = new Set<string>();
-    const participantFiles = new Set<string>();
-    const lcaPrefix = lca === "" ? "" : lca + "/";
-    for (const p of participants) {
-      if (!p.startsWith(lcaPrefix) && p !== lca) continue;
-      const rest = p === lca ? "" : p.slice(lcaPrefix.length);
-      const firstSegment = rest.split("/")[0];
-      const container = lca === "" ? firstSegment : `${lca}/${firstSegment}`;
-      expandedContainers.add(container);
-      // Only add as a participant file if it's actually a file node
-      // (an edge endpoint may be a package path on some samples).
-      const node = data.nodes.find((n) => n.id === p);
-      if (node?.kind === "file") participantFiles.add(p);
-    }
-    if (lca !== state.scope) state.setState({ scope: lca });
-    setIssueFocus({
-      key,
-      lca,
-      edges,
-      expandedContainers,
-      participantFiles,
-    });
+    const focus = computeIssueFocus(issue, key, data);
+    if (focus.lca !== state.scope) state.setState({ scope: focus.lca });
+    setIssueFocus(focus);
   }
 
   function onNodePicked(id: string) {
@@ -259,58 +204,6 @@ export function App({ data }: AppProps) {
       />
     </div>
   );
-}
-
-/**
- * Extract the participant file ids from an issue. Cycles enumerate
- * their full member list; red flags carry an explicit `modules` array
- * (passed through via the Issue.scope). Violations and gravity wells
- * carry a single subject.
- */
-function participantsOf(it: Issue, data: DataContract): string[] {
-  if (it.kind === "cycle") {
-    const c = data.cycles.find((c) => c.members[0] === it.subject);
-    return c?.members ?? [it.subject];
-  }
-  if (it.kind === "red-flag") {
-    const f = data.red_flags.find((f) => f.modules[0] === it.subject);
-    return f?.modules ?? [it.subject];
-  }
-  if (it.kind === "violation") {
-    const v = data.violations.find(
-      (v) => v.edge.from === it.subject || v.edge.to === it.subject,
-    );
-    return v ? [v.edge.from, v.edge.to] : [it.subject];
-  }
-  return [it.subject];
-}
-
-/**
- * Longest common ancestor of a set of paths. Splits each path into
- * `/`-segments and keeps prefix segments that all paths agree on.
- * Returns "" when there's nothing in common (drilling to root).
- */
-function longestCommonAncestor(paths: string[]): string {
-  if (paths.length === 0) return "";
-  const splits = paths.map((p) => p.split("/").filter(Boolean));
-  const common: string[] = [];
-  for (let i = 0; ; i++) {
-    const seg = splits[0][i];
-    if (seg === undefined) break;
-    if (!splits.every((s) => s[i] === seg)) break;
-    common.push(seg);
-  }
-  // Drop the last segment if it's a leaf (file) — we want the
-  // containing directory, not the file itself.
-  if (
-    common.length > 0 &&
-    paths.length === 1 &&
-    common.join("/") === paths[0] &&
-    paths[0].includes(".")
-  ) {
-    common.pop();
-  }
-  return common.join("/");
 }
 
 function codebaseTitleOf(data: DataContract): string {
